+++ www/api/libs/domain_entity/__init__.py
@@ -1,0 +1,1 @@
+from .flask_file import FlaskFile
