+++ www/api/libs/domain_entity/flask_file.py
