+++ www/api/libs/utils/__init__.py
@@ -1,4 +1,3 @@
 from .api_response import api_response
-from .db_connection import get_db_engine
 from .sqlmodel_utils import as_dict, as_json_dumps, TrackTimeMixin, SoftDeleteMixin, BaseModel
 from .validate_dict import validate_dict, VKOptions
