+++ www/api/main.py
@@ -1,13 +1,7 @@
 from flask import Blueprint
 
-from api.modules.video.infraestructure.controller import video_api
+from api.modules.video.infraestructure.controllers import video_api
 
 api = Blueprint('api', __name__)
-
-
-@api.route('/')
-def index():
-    return 'OK'
-
 
 api.register_blueprint(video_api, url_prefix='/video')
