+++ www/api/modules/s3/infraestructure/repository/minio_s3_repository.py
@@ -3,7 +3,7 @@ import minio
 
 from minio.error import S3Error
 
-from api.shared.domain import FlaskFile
+from api.libs.domain_entity import FlaskFile
 
 
 class MinioS3Repository:
