+++ www/api/modules/video/infraestructure/controllers/video_http_controller.py
@@ -1,8 +1,7 @@
 import json
 from flask import Blueprint, request
 
-from api.shared.infrastructure.utils import (
-    api_response as Response, as_json_dumps)
+from api.libs.utils import api_response as Response, as_json_dumps
 
 from ... import application
