--- www/api/shared/domain/__init__.py
@@ -1,1 +1,0 @@
-from .entity import FlaskFile
