--- www/api/shared/domain/entity/__init__.py
@@ -1,1 +1,0 @@
-from .flask_file import FlaskFile
