--- www/api/shared/infrastructure/controller/__init__.py
@@ -1,1 +1,0 @@
-from .api_controller import api
